Rename @Token to @UseSessions
